move all_estimators tests
