+++ sktime/registry/tests/test_lookup.py
@@ -111,3 +111,46 @@ def test_all_tags(estimator_scitype):
             assert isinstance(tag[2][0], str)
             assert isinstance(tag[2][1], list)
         assert isinstance(tag[3], str)
+
+
+@pytest.mark.parametrize("return_names", [True, False])
+def test_all_estimators_return_names(return_names):
+    """Test return_names argument in all_estimators."""
+    estimators = all_estimators(return_names=return_names)
+    assert isinstance(estimators, list)
+    assert len(estimators) > 0
+
+    if return_names:
+        assert all([isinstance(estimator, tuple) for estimator in estimators])
+        names, estimators = list(zip(*estimators))
+        assert all([isinstance(name, str) for name in names])
+        assert all(
+            [name == estimator.__name__ for name, estimator in zip(names, estimators)]
+        )
+
+    assert all([isinstance(estimator, type) for estimator in estimators])
+
+
+# arbitrary list for exclude_estimators argument test
+EXCLUDE_ESTIMATORS = [
+    "ElasticEnsemble",
+    "ProximityForest",
+    "ProximityStump",
+    "ProximityTree",
+]
+
+
+@pytest.mark.parametrize("exclude_estimators", ["NaiveForecaster", EXCLUDE_ESTIMATORS])
+def test_all_estimators_exclude_estimators(exclude_estimators):
+    """Test exclued_estimators argument in all_estimators."""
+    estimators = all_estimators(
+        return_names=True, exclude_estimators=exclude_estimators
+    )
+    assert isinstance(estimators, list)
+    assert len(estimators) > 0
+    names, estimators = list(zip(*estimators))
+
+    if not isinstance(exclude_estimators, list):
+        exclude_estimators = [exclude_estimators]
+    for estimator in exclude_estimators:
+        assert estimator not in names
